+++ Magic/src/resource-pack/build.sh
@@ -6,33 +6,27 @@ cd target
 
 # New resource packs
 
-echo "** BUILDING SURVIVAL **"
+echo "** BUILDING DEFAULT **"
 
-mkdir survival
-cd survival
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../survival/* .
+mkdir default
+cd default
+cp -R ../../default/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-RP-6-2.zip *
+zip -r -X ../Magic-RP-6-3.zip *
 cd ..
 
-echo "** BUILDING SURVIVAL-Skulls **"
+echo "** BUILDING DEFAULT-Skulls **"
 
-mkdir survival-skulls
-cd survival-skulls
+mkdir default-skulls
+cd default-skulls
 mkdir assets
 cd assets
-cp -R ../../../common/assets/* .
 cp -R ../../../skulls/assets/* .
 cd ..
-cp -R ../../survival/* .
+cp -R ../../default/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-skulls-RP-6-2.zip *
+zip -r -X ../Magic-skulls-RP-6-3.zip *
 cd ..
-
 
 echo "** BUILDING SKULLS **"
 
@@ -48,67 +42,8 @@ echo "** BUILDING PAINTERLY **"
 
 mkdir painterly
 cd painterly
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../survival/* .
+cp -R ../../default/* .
 cp -R ../../painterly/* .
 find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-painterly-RP-6-2.zip *
-cd ..
-
-echo "** BUILDING POTTER **"
-
-mkdir potter 
-cd potter
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../potter/* .
-find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-potter-RP-6.zip *
-cd ..
-
-echo "** BUILDING BENDING **"
-
-mkdir bending
-cd bending
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../bending/* .
-find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-bending-RP-6.zip *
-cd ..
-
-echo "** BUILDING STARS **"
-
-mkdir stars
-cd stars
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../stars/* .
-find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-stars-RP-6.zip *
-cd ..
-
-echo "** BUILDING FFA **"
-
-mkdir ffa
-cd ffa
-mkdir assets
-cd assets
-cp -R ../../../common/assets/* .
-cd ..
-cp -R ../../bending/* .
-cp -R ../../potter/* .
-cp -R ../../stars/* .
-cp -R ../../survival/* .
-find . -name ".DS_Store" -type f -delete
-zip -r -X ../Magic-ffa-RP-6.zip *
+zip -r -X ../Magic-painterly-RP-6-3.zip *
 cd ..